This commit tries to eliminate multiple interpreters (for animation, sequence diagrams, etc) As well as people needing to know about 'foldMap'.
To do that, I introduced 'RunnableProgram' as a trait where you could just "insert here" the 'how'
of how run the logic.

That 'how' function returns a Result, which can either be a Task (just something to do in-process),
or else a TraceTask (something we want to track as we communicate with what would be other services).

We refer to systems as 'Coords' (a namespace and a service name).
The Results then let us track what services are talking to each other in our telemetry.

After running out business logic, we can ask the Telemetry for a call stack, and then turn that
call stack into a sequence diagram, animation, etc.

There is an implicit conversion from a Task[A] to a Result[A]. We should also be able to pimp a Task
for perhaps better readability as:

someTask.asResult
someTask.asTracedResult(targetService : Coords)

* Added openapi task to sbt, refreshed the UI
* started complex restaurant logic
* added 'RunnableProgram' and Telemetry